requirements.txt
requests==2.32.3
asyncpg==0.30.0
python-dotenv==1.1.1
pip==24.0
asyncio==3.4.3
aiogram==3.20.0
APScheduler==3.11.0